Restructure file locations
