+++ hen-ui/src/pages/createParcel/tabs/SenderContactsTab.tsx
@@ -1,38 +1,21 @@
 import React from 'react';
-import { Text, List, Button } from '../../../components';
-import DeliveryMethodCard from './DeliveryMethodCard';
-import deliveryMethodsMock from '../../../mocks/deliveryMethodsMock';
+import { Button, Text } from '../../../components';
+import ContactInformationForm from '../components/ContactInformationForm';
 import { ArrowSVG } from '../../../assets/images/arrow';
 
-function DeliveryMethodsTab() {
-  const [activeCard, setActiveCard] = React.useState<string | undefined>(
-    undefined,
-  );
-
+function SenderContactsTab() {
   return (
-    <div className="flex flex-col items-center justify-start w-full pt-6 pr-10 pl-8">
-      <div className="flex flex-col gap-[21px] items-start justify-start w-full">
+    <div className="flex flex-row gap-[25px] justify-start mt-[7px] w-full">
+      <div className="flex flex-col items-start justify-between md:ml-[0] ml-8 pr-10 md:w-full mt-5">
         <Text
-          className="ml-0.5 md:ml-[0] text-indigo_600 text-left w-auto"
+          className="text-left text-indigo_600 mb-4"
           as="h2"
           variant="h2"
         >
-          Choose your delivery type:
+          Your contact information
         </Text>
-        <List
-          className="sm:flex-col flex-row md:gap-10 gap-10 grid sm:grid-cols-1 md:grid-cols-2 grid-cols-3 justify-center w-full"
-          orientation="horizontal"
-        >
-          {deliveryMethodsMock.map((deliveryMethod, index) => (
-            <DeliveryMethodCard
-              key={index}
-              {...deliveryMethod}
-              isActive={activeCard === deliveryMethod.deliveryType}
-              onClick={() => setActiveCard(deliveryMethod.deliveryType)}
-            />
-          ))}
-        </List>
-        <div className="flex flex-row gap-10 w-full justify-between mt-28">
+        <ContactInformationForm />
+        <div className="flex flex-row gap-10 w-full justify-between">
           <Button
             className="flex flex-row gap-2 justify-center items-center cursor-pointer font-medium leading-[normal] min-w-[190px] mt-5 text-center text-md text-indigo_600 w-auto"
             shape="RoundedBorder15"
@@ -73,4 +56,4 @@ function DeliveryMethodsTab() {
   );
 }
 
-export default DeliveryMethodsTab;
+export default SenderContactsTab;
